requirements.txt
DateTime==4.0.1
funcsigs==0.4
matplotlib==1.4.3
mock==1.2.0
nose==1.3.7
numpy==1.9.2
pbr==1.3.0
pyparsing==2.0.3
pyserial==2.7
python-dateutil==2.4.2
pytz==2015.4
six==1.9.0
wheel==0.24.0
zope.interface==4.1.2
